JavaScrip код проекта разбит на 3 модуля: data, util и main. Где последний является точкой входа.
